Move Settings in the effects module
