--- turbo_audio/src/resources/effects/mod.rs
@@ -1,10 +1,0 @@
-use self::{lua::LuaEffect, native::NativeEffect};
-
-pub mod lua;
-pub mod native;
-
-#[derive(Debug)]
-pub enum Effect {
-    Lua(LuaEffect),
-    Native(NativeEffect),
-}
