+++ turbo_audio/src/resources/mod.rs
@@ -1,4 +1,2 @@
 pub mod color;
-pub mod effects;
 pub mod ledstrip;
-pub mod settings;
