--- turbo_audio/src/resources/settings.rs
@@ -1,7 +1,0 @@
-use super::effects::{lua::LuaEffectSettings, native::NativeEffectSettings};
-
-#[derive(Debug)]
-pub enum Settings {
-    Lua(LuaEffectSettings),
-    Native(NativeEffectSettings),
-}
